refactor : 검수 대기 목록 타입 shared로 분리
* 검수 대기 목록 타입 4종(상태·항목·페이지네이션·목록)을 partner/_shared/model로 이동
* review 세그먼트·파트너 대시보드의 목록 타입 참조를 shared 경로로 변경
* review 타입에는 검수 현황 요약 타입만 잔류
