+++ apps/web/src/app/partner/_components/PendingReviewPreview.tsx
@@ -5,7 +5,7 @@ import { StatusBadge } from "@/shared/ui/StatusBadge";
 import { buttonVariants } from "@/shared/ui/Button";
 import { accidentTypeLabel } from "@/shared/model/accident-type";
 import { useReviewList } from "../review/_api/use-review-list";
-import type { ReviewListItem } from "../review/_model/types";
+import type { ReviewListItem } from "../_shared/model/types";
 import { SectionCard } from "./SectionCard";
 import { PendingReviewEmpty } from "./PendingReviewEmpty";
