+++ apps/web/src/app/partner/_shared/model/types.ts
@@ -1,0 +1,12 @@
+import type { z } from "zod";
+import type {
+  reviewStatusSchema,
+  reviewListItemSchema,
+  paginationSchema,
+  reviewListSchema,
+} from "./review-list.schema";
+
+export type ReviewStatus = z.infer<typeof reviewStatusSchema>;
+export type ReviewListItem = z.infer<typeof reviewListItemSchema>;
+export type Pagination = z.infer<typeof paginationSchema>;
+export type ReviewList = z.infer<typeof reviewListSchema>;
