+++ apps/web/src/app/partner/review/_api/get-review-list.ts
@@ -2,7 +2,7 @@ import { API_BASE_URL } from "@/shared/api/config";
 import { fetchJson } from "@/shared/api/fetch-json";
 import type { ReviewListFilter } from "@/shared/api/query-keys";
 import { reviewListSchema } from "../../_shared/model/review-list.schema";
-import type { ReviewList } from "../_model/types";
+import type { ReviewList } from "../../_shared/model/types";
 
 export function getReviewList(filter?: ReviewListFilter): Promise<ReviewList> {
   const params = new URLSearchParams();
